Generalised SpigotConfigManager to YamlConfigManager
