+++ src/main/java/me/jsedwards/SpigotConfigManager.java
@@ -2,29 +2,26 @@ package me.jsedwards;
 
 import me.jsedwards.gui.Server;
 import org.apache.commons.lang3.StringUtils;
-import org.apache.logging.log4j.LogManager;
-import org.apache.logging.log4j.Logger;
 import org.jsoup.Jsoup;
 import org.jsoup.nodes.Document;
 import org.jsoup.nodes.Element;
 import org.jsoup.nodes.Node;
 import org.jsoup.nodes.TextNode;
-import org.yaml.snakeyaml.Yaml;
 
-import javax.swing.*;
-import java.io.*;
-import java.nio.file.Files;
-import java.util.*;
+import java.io.File;
+import java.io.IOException;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Optional;
 import java.util.regex.MatchResult;
 import java.util.regex.Pattern;
 
-public class SpigotConfigManager extends DefaultListModel<String> implements ConfigManager {
-
-    private static final Logger LOGGER = LogManager.getLogger();
+public class SpigotConfigManager extends YamlConfigManager {
 
     private static final HashMap<String, String> PROPERTY_DESCRIPTIONS = new HashMap<>();
     private static final HashMap<String, String> PROPERTY_DATA_TYPES = new HashMap<>();
     private static final HashMap<String, String> PROPERTY_DEFAULTS = new HashMap<>();
+
     static {
         Pattern pattern = Pattern.compile("Default: ((?:.(?!Type:))*) Type: ((?:.(?!Description:))*) Description: ((?:.(?!Default:))*)");
         try {
@@ -66,99 +63,8 @@ public class SpigotConfigManager extends DefaultListModel<String> implements Con
         }
     }
 
-    private final File yamlFile;
-    private Map<String, Object> map = new HashMap<>();
-    private final List<String> keys = new ArrayList<>();
-    private List<String> filteredKeys;
-    private boolean saved = true;
-
     public SpigotConfigManager(Server server) {
-        if (server == null) {
-            yamlFile = null;
-        } else {
-            yamlFile = new File(server.serverLocation + File.separator + "spigot.yml");
-            Yaml yaml = new Yaml();
-            if (Files.exists(yamlFile.toPath())) {
-                try (InputStream stream = new FileInputStream(yamlFile)) {
-                    map = yaml.load(stream);
-                    LOGGER.info("Loaded spigot config from " + yamlFile.getAbsolutePath());
-                } catch (IOException e) {
-                    LOGGER.error("Failed to load spigot config from " + yamlFile.getAbsolutePath(), e);
-                }
-            }
-        }
-        explore(map, "");
-        filteredKeys = new ArrayList<>(keys);
-        LOGGER.info("Loaded spigot config");
-    }
-
-    private void explore(Map<?, ?> in, String current) {
-        for (Object key : in.keySet()) {
-            String newCurrent = current + "/" + key;
-            Object value = in.get(key);
-            if (value instanceof Map) {
-                explore((Map<?, ?>) value, newCurrent);
-            } else {
-                keys.add(newCurrent.substring(1));
-            }
-        }
-    }
-
-    private Object getFromPath(String path) {
-        Object next = map;
-        String[] split = path.split("/");
-        for (String s : split) {
-            next = ((Map<?, ?>) next).get(s);
-        }
-        return next;
-    }
-
-    @Override
-    public void updateSearch(String query) {
-        filteredKeys = keys.stream().filter(s -> s.contains(query)).toList();
-    }
-
-    @Override
-    public void save() {
-        if (!saved) realSave();
-    }
-
-    private void realSave() {
-        try (BufferedWriter writer = new BufferedWriter(new FileWriter(yamlFile))) {
-            Yaml yaml = new Yaml();
-            String dump = yaml.dumpAsMap(map);
-            writer.write(dump);
-            saved = true;
-        } catch (IOException e) {
-            LOGGER.error("Failed to save spigot config to " + yamlFile.getAbsolutePath(), e);
-        }
-    }
-
-    @Override
-    public void set(String key, String value) {
-        Object next = map;
-        Object last = null;
-        String[] split = key.split("/");
-        for (String s : split) {
-            last = next;
-            next = ((Map<?, ?>) next).get(s);
-        }
-        Map<Object, Object> map = (Map<Object, Object>) last;
-        String lastKey = split[split.length - 1];
-        Object current = map.get(lastKey);
-        Object castValue = cast(value, current.getClass());
-        map.put(lastKey, castValue);
-        saved = false;
-    }
-
-    private Object cast(String value, Class<?> clazz) {
-        if (clazz == String.class) return value;
-        if (clazz == Integer.class) return Integer.parseInt(value);
-        if (clazz == Boolean.class) return Boolean.valueOf(value);
-        if (clazz == Double.class) return Double.parseDouble(value);
-        if (clazz == Float.class) return Float.parseFloat(value);
-        LOGGER.warn("Unsupported class found in %s: %s".formatted(yamlFile.getAbsolutePath(), clazz.getName()));
-        return value;
+        super(server, s -> s.serverLocation + File.separator + "spigot.yml");
     }
 
     @Override
@@ -179,16 +85,5 @@ public class SpigotConfigManager extends DefaultListModel<String> implements Con
     @Override
     public void optimise(int sliderValue) {
 
-    }
-
-    @Override
-    public int getSize() {
-        return filteredKeys.size();
-    }
-
-    @Override
-    public String getElementAt(int index) {
-        String key = filteredKeys.get(index);
-        return key + ": " + getFromPath(key);
     }
 }
